move into dirs
