+++ selfdrive/loggerd/SConscript
@@ -5,11 +5,11 @@ libs = [common, cereal, messaging, visionipc,
         'avformat', 'avcodec', 'swscale', 'avutil',
         'yuv', 'OpenCL', 'pthread']
 
-src = ['logger.cc', 'loggerd.cc', 'video_writer.cc', 'remote_encoder.cc', 'encoder.cc']
+src = ['logger.cc', 'loggerd.cc', 'encoder/video_writer.cc', 'remote_encoder.cc', 'encoder/encoder.cc']
 if arch == "larch64":
-  src += ['v4l_encoder.cc']
+  src += ['encoder/v4l_encoder.cc']
 else:
-  src += ['raw_logger.cc']
+  src += ['encoder/raw_logger.cc']
 
 if arch == "Darwin":
   # fix OpenCL
