+++ selfdrive/loggerd/encoder/video_writer.h
